pyproject.toml
[project]
name = "llm-lift-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.1.4",
    "markdown>=3.7",
    "mediapipe>=0.10.21",
    "openai>=1.64.0",
    "opencv-python>=4.11.0.86",
    "pypandoc>=1.15",
    "python-dotenv>=1.0.1",
    "reportlab>=4.3.1",
    "streamlit>=1.42.2",
    "streamlit-image-coordinates>=0.2.0",
    "weasyprint>=64.1",
]
